# Conflicts:
#	src/main/java/sausageShopBack/controllers/RegisterController.java
#	src/main/java/sausageShopBack/models/Product.java
#	src/main/resources/templates/login.html
#	src/main/resources/templates/myaccount.html
#	src/main/resources/templates/register.html

+++ src/main/java/sausageShopBack/models/Category.java
@@ -29,7 +29,7 @@ public class Category implements Serializable, Comparable<Category> {
     @NotNull
     @GeneratedValue
     @Column(name = "id")
-    private Integer id = null;
+    private Long id = null;
 
     public Category(@NotNull String title) {
         this.products = new ArrayList<>();
